Conflicts:
	composer.json
	composer.lock

+++ application/orga/controllers/Datagrid/GranularityController.php
@@ -80,7 +80,7 @@ class Orga_Datagrid_GranularityController extends UI_Controller_Datagrid
     {
         $organization = Orga_Model_Organization::load($this->getParam('idOrganization'));
 
-        $refAxes = explode(',', $this->getAddElementValue('axes'));
+        $refAxes = $this->getAddElementValue('axes');
         $listAxes = array();
         $refGranularity = '';
         if (empty($refAxes)) {
@@ -119,7 +119,8 @@ class Orga_Datagrid_GranularityController extends UI_Controller_Datagrid
                     (bool) $this->getAddElementValue('dW'),
                     (bool) $this->getAddElementValue('genericActions'),
                     (bool) $this->getAddElementValue('contextActions'),
-                    (bool) $this->getAddElementValue('inputDocuments')
+                    (bool) $this->getAddElementValue('inputDocuments'),
+                    __('Orga', 'backgroundTasks', 'addGranularity', ['LABEL' => implode(', ', $listAxes)])
                 )
             );
             $this->message = __('UI', 'message', 'addedLater');
@@ -144,6 +145,9 @@ class Orga_Datagrid_GranularityController extends UI_Controller_Datagrid
     public function deleteelementAction()
     {
         $granularity = Orga_Model_Granularity::load($this->delete);
+        if ($granularity->getCellsWithACL()) {
+            throw new Core_Exception_User('Orga', 'granularity', 'granularityCantBeDeleted');
+        }
 
         $granularity->delete();
 
@@ -191,6 +195,14 @@ class Orga_Datagrid_GranularityController extends UI_Controller_Datagrid
                 $this->data = $granularity->getCellsWithOrgaTab();
                 break;
             case 'aCL':
+                foreach ($granularity->getCells() as $cell) {
+                    $cellResource = User_Model_Resource_Entity::loadByEntity($cell);
+                    foreach ($cellResource->getLinkedSecurityIdentities() as $linkedIdentity) {
+                        if (!($linkedIdentity instanceof User_Model_Role) || (count($linkedIdentity->getUsers()) > 0)) {
+                            throw new Core_Exception_User('Orga', 'granularity', 'roleExistsForCellAtThisGranularity');
+                        }
+                    }
+                }
                 $granularity->setCellsWithACL((bool) $this->update['value']);
                 $this->data = $granularity->getCellsWithACL();
                 break;
